Demopro/requirements.txt
Flask==2.0.3
pytesseract==0.3.8
Pillow==8.4.0
python-3.10.0
